+++ src/design/plone/contenttypes/restapi/serializers/dxfields.py
@@ -1,6 +1,7 @@
 # -*- coding: utf-8 -*-
 from AccessControl.unauthorized import Unauthorized
 from Acquisition import aq_inner
+from collective.volto.enhancedlinks.interfaces import IEnhancedLinksEnabled
 from design.plone.contenttypes.interfaces import IDesignPloneContenttypesLayer
 from design.plone.contenttypes.interfaces.servizio import IServizio
 from plone import api
@@ -8,12 +9,16 @@ from plone.app.contenttypes.utils import replace_link_variables_by_paths
 from plone.base.utils import human_readable_size
 from plone.dexterity.interfaces import IDexterityContent
 from plone.namedfile.interfaces import INamedFileField
+from plone.namedfile.interfaces import INamedImageField
 from plone.outputfilters.browser.resolveuid import uuidToURL
 from plone.restapi.interfaces import IBlockFieldSerializationTransformer
 from plone.restapi.interfaces import IFieldSerializer
 from plone.restapi.interfaces import ISerializeToJsonSummary
 from plone.restapi.serializer.converters import json_compatible
 from plone.restapi.serializer.dxfields import DefaultFieldSerializer
+from plone.restapi.serializer.dxfields import (
+    ImageFieldSerializer as BaseImageFieldSerializer,
+)
 from zope.component import adapter
 from zope.component import getMultiAdapter
 from zope.component import subscribers
@@ -22,6 +27,7 @@ from zope.interface import implementer
 from zope.schema.interfaces import IList
 from zope.schema.interfaces import ISourceText
 from zope.schema.interfaces import ITextLine
+
 
 import json
 import re
@@ -80,9 +86,15 @@ class FileFieldViewModeSerializer(DefaultFieldSerializer):
             "filename": namedfile.filename,
             "content-type": namedfile.contentType,
             "size": size,
-            "getObjSize": human_readable_size(size),
             "download": url,
         }
+        if IEnhancedLinksEnabled.providedBy(self.context):
+            result.update(
+                {
+                    "getObjSize": human_readable_size(size),
+                    "enhanced_links_enabled": True,
+                }
+            )
 
         return json_compatible(result)
 
@@ -93,6 +105,20 @@ class FileFieldViewModeSerializer(DefaultFieldSerializer):
                 return "@@display-file"
 
         return "@@download"
+
+
+@adapter(INamedImageField, IDexterityContent, IDesignPloneContenttypesLayer)
+class ImageFieldSerializer(BaseImageFieldSerializer):
+    def __call__(self):
+        result = super().__call__()
+        if result and IEnhancedLinksEnabled.providedBy(self.context):
+            result.update(
+                {
+                    "getObjSize": human_readable_size(result["size"]),
+                    "enhanced_links_enabled": True,
+                }
+            )
+        return result
 
 
 def serialize_data(context, json_data, show_children=False):